+++ Lib/Exported_Functions.js
@@ -1,0 +1,19 @@
+import { ModuleName, ModuleTitle } from "../ss-initialize.js";
+import { log } from "./Lib/Exported_Functions.js";
+
+/** A wrapper function that works with the Foundryvtt-devMode module to output debugging info
+ *  to the console.log, when a debugging boolean is activated in module settings.
+ *  Or the code can pass TRUE to the force argument to output to console.log regardless of the debugging boolean.
+ *  @param {Boolean}    force   - A manual bypass to force output regardless of the debugging boolean
+ *  @param {}           args    - The content to be output to console.log
+ *  @return {void}
+ */
+ export function log(force, ...args) {
+    const isDebugging = game.modules.get('_dev-mode')?.api?.getPackageDebugValue(ModuleName);
+
+    if ( isDebugging ) {
+        console.log(ModuleTitle,  "DEBUG | ", ...args);
+    } else if ( force ) {
+        console.log(ModuleTitle, " | ", ...args)
+    }
+}
